Refactors nodal calc in reflector ADF generation. Can reuse nodal solver elsewhere now.
